+++ src/leapmotor_api/mappings.py
@@ -14,6 +14,7 @@ from .const import (
     REMOTE_CTL_BATTERY_PREHEAT_ON,
     REMOTE_CTL_BLE_KEY_RESTART,
     REMOTE_CTL_CHARGE_LIMIT,
+    REMOTE_CTL_CHARGE_SCHEDULE,
     REMOTE_CTL_CHARGE_START,
     REMOTE_CTL_CHARGE_STOP,
     REMOTE_CTL_FIND_CAR,
@@ -192,6 +193,7 @@ REMOTE_ACTION_SPECS: dict[str, RemoteActionSpec] = {
         required_right=VehicleRight.WINDSHIELD_DEFROST,
     ),
     REMOTE_CTL_CHARGE_LIMIT: RemoteActionCtlChargePlan(required_right=VehicleRight.CHARGE_LIMIT),
+    REMOTE_CTL_CHARGE_SCHEDULE: RemoteActionCtlChargePlan(required_right=VehicleRight.CHARGE_LIMIT),
     REMOTE_CTL_SEND_DESTINATION: RemoteActionCtlSendDestination(required_right=VehicleRight.SEND_DESTINATION),
     REMOTE_CTL_SENTRY_MODE: RemoteActionCtlSentryMode(
         value=SentryModeValue.ON, required_right=VehicleRight.SENTRY_MODE